requirements.txt
streamlit
numpy
sounddevice
librosa
google-generativeai
python-dotenv
